+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.18](https://github.com/nicklockwood/SwiftFormat/releases/tag/0.18) (2016-11-??)
 
 - Added `--ifdef` command line option for controlling how `#if`...`#endif` clauses are indented
+- Added `redundantGet` rule to remove unneeded `get {}` clause in read-only computed properties
 - `redundantParens` now removed unneeded parens from expressions and closure arguments
 
 ## [0.17.2](https://github.com/nicklockwood/SwiftFormat/releases/tag/0.17.2) (2016-11-11)